+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Fit `CrowdfundingCard` image to new 16/10 ratio.
+- Feature: Add `stretch` prop to `CrowdfundingCard`.
+
 ## [2.65.0] - 2020-04-21
 
 Doc: